% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cleanMetadata.GSE10846}
\alias{cleanMetadata.GSE10846}
\title{Specific function for cleaning GSE10846}
\usage{
\method{cleanMetadata}{GSE10846}(meta_data)
}
\arguments{
\item{meta_data}{A \code{data.frame} as returned by
\code{\link{downloadAndPreparemeta_data}}.}

\item{A}{clean \code{data.frame}.}
}
\description{
Specific function for cleaning GSE10846
}

