% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cleanMetadata.GSE31312}
\alias{cleanMetadata}
\alias{cleanMetadata.GSE31312}
\alias{cleanMetadata.data.frame}
\title{Generic metadata cleaning function}
\usage{
\method{cleanMetadata}{GSE31312}(meta_data)

cleanMetadata(meta_data)

\method{cleanMetadata}{data.frame}(meta_data)
}
\arguments{
\item{meta_data}{Output from \code{downloadAndPrepareMetadata}.}
}
\value{
Returns a function that will clean the given GEO dataset
}
\description{
S3 method dispatch giving the custom functions for formatting datasets.
The function \code{downloadAndPrepareMetadata} adds a class \code{"GSEXXXX"}
of the GEO number in question.
The generic looks for functions of the name \code{cleanMetadata.GSEXXXX} and
specific for that dataset and cleans metadata using that function.
}
\details{
GSE31312:\cr
   This function makes the samples GSM776068 and GSM776149 to be left
   out of the downsteam preprocessing due to bad array quality.
}
\author{
Anders Ellern Bilgrau, Steffen Falgreen Larsen
}

