% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadAndPrepareCELFiles}
\alias{downloadAndPrepareCELFiles}
\title{Download GEO dataset}
\usage{
downloadAndPrepareCELFiles(geo_nbr, destdir = getwd(), clean = TRUE,
  verbose = TRUE)
}
\arguments{
\item{geo_nbr}{The GEO ascession number.}

\item{destdir}{The destination dir of the downloaded files.}

\item{clean}{Should the strictly unnessesary files be deleted?}

\item{verbose}{Signal the process.}
}
\description{
Function for downloading a GEO dataset and prepare it for preprocessing.
The function uses \code{getGEOSuppFiles} which downloads the data. The
function then unzips the data files in the GEO series.
}
\note{
The function will overwrite existing files in the \code{destdir}.
}
\examples{
\dontrun{
downloadAndPrepareCELFiles(geo_nbr = "GSE18376",
                          destdir = tempdir())
}
}
\author{
Anders Ellern Bilgrau,
  Steffen Falgreen Larsen
}

