% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadAndPrepareGEOData}
\alias{downloadAndPrepareGEOData}
\title{Download GEO dataset}
\usage{
downloadAndPrepareGEOData(geo_nbr, destdir = getwd(), verbose = FALSE,
  clean = TRUE)
}
\arguments{
\item{geo_nbr}{The GEO ascession number.}

\item{destdir}{The destination dir of the downloaded files.}

\item{verbose}{Signal the process.}

\item{clean}{Should the strictly unnessesary files be deleted?}
}
\description{
Function for downloading a GEO dataset. Simply a wrapper for \code{getGEO}
and \code{getGEOSuppFiles}.
}

