% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadAndPrepareMetadata}
\alias{downloadAndPrepareMetadata}
\title{Download GEO metadata}
\usage{
downloadAndPrepareMetadata(geo_nbr, destdir = getwd(), clean = FALSE,
  verbose = TRUE)
}
\arguments{
\item{geo_nbr}{The GEO ascession number.}

\item{destdir}{The destination dir of the downloaded files.}

\item{clean}{Should the strictly unnessesary files be deleted?}

\item{verbose}{Signal the process.}
}
\value{
A \code{data.frame} giving the clinical and metadata information
  for the GEO dataset.
}
\description{
Function for downloading the metadata of a GEO dataset and prepare it for
analysis.
The function uses \code{getGEOSuppFiles} which downloads the data. The
function then unzips the data files in the GEO series.
}
\note{
The function will overwrite existing files in the \code{destdir}.
}
\examples{
\dontrun{
downloadAndPrepareCELFiles(geo_nbr = "GSE18376",
                           destdir = tempdir())
}
}
\author{
Anders Ellern Bilgrau,
  Steffen Falgreen Larsen
}
\keyword{internal}

