% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLatestVersion}
\alias{getLatestVersion}
\title{Get latest brainarray version}
\usage{
getLatestVersion()
}
\value{
Returns a string on the form X.Y.Z giving the latest version.
}
\description{
Currently a very hacky implementation. To be updated later. The function
currently reads the tables available at
\url{http://brainarray.mbni.med.umich.edu/Brainarray/Database/CustomCDF/CDF_download.asp}
to determine the latest version
}
\examples{
getLatestVersion()
}

