% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listTargets.R
\name{listTargets}
\alias{listTargets}
\title{List available targets}
\usage{
listTargets(cdf, version = getLatestVersion())
}
\arguments{
\item{cdf}{A character. Either \code{"affy"} or \code{"brainarray"}.}

\item{version}{A character giving the brainarray version.
See \code{listVersions()}. Ignored if \code{cdf == "affy"}.}
}
\value{
A character vector of the available targets.
}
\description{
This function lists available targets for a given \code{cdf} argument.
}
\examples{
listTargets("affy")
listTargets("brainarray", version = "18.0.0")
listTargets("brainarray", version = "17.1.0")
listTargets("brainarray", version = "15.0.0")
}

