% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary_brainarray.R
\name{listVersions}
\alias{listVersions}
\title{Available Brainaaray versions}
\usage{
listVersions()
}
\value{
A character vector of the available versions.
}
\description{
Lists the available Brainaaray versions for \pkg{DLBCLdata}. The list is
currently hard coded. Newer versions may work.
}

