% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readBrainarrayTable.R
\name{readBrainarrayTable}
\alias{readBrainarrayTable}
\title{Reads table from brainarray homepage}
\usage{
readBrainarrayTable(custom_cdf, version = getLatestVersion())
}
\arguments{
\item{custom_cdf}{A string giving the "target" to use.}

\item{version}{A string giving the version.}
}
\value{
A \code{data.frame} of the info on the relevant brainarray webpage.
}
\description{
Reads table from brainarray homepage
}
\examples{
readBrainarrayTable("ENSG", version = "17.0.0")[1:10, 1:10]
readBrainarrayTable("ENST", version = "18.0.0")[1:10, 1:10]
readBrainarrayTable("ENSE", version = "19.0.0")[1:10, 1:10]
readBrainarrayTable("ENSG")[1:10, 1:10]
}
\keyword{internal}

