% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readDACData}
\alias{readDACData}
\title{Read DLBCL class files}
\usage{
readDACData(geo_nbr)
}
\arguments{
\item{geo_nbr}{A character string giving the GSE number.}
}
\value{
A \code{data.frame} of the predicted classes and the class
  probabilities.
}
\description{
Functions to read the internal data files of classified CEL files using
the DLBCL Automatic Classifier (DAC).
}
\examples{
readDACData("GSE19246")
}
\keyword{internal}

