% Generated by roxygen2 (4.0.2): do not edit by hand
\name{requireBrainarray}
\alias{requireBrainarray}
\title{Installs brainarray CDF and probe information packages}
\usage{
requireBrainarray(array_type, custom_cdf = "enst",
  version = getLatestVersion())
}
\arguments{
\item{array_type}{A character of length 1 giving the giving the name of the'
array. (Preferably as returned by \code{affyio::read.celfile.header})}

\item{custom_cdf}{A charcter giving the annotation type.}

\item{version}{A character of length 1 giving the version to download.}
}
\value{
Invisibly returns a list of the locations of the saved files.
}
\description{
This package downloads, installs, and/or loads the brainarray CDF and probe
enviroments from the Brainarray website.
}
\examples{
tmp.path <- requireBrainarray("hgu133a", custom_cdf = "enst")
print(tmp.path)
}
\author{
Steffen Falgreen
  Anders Ellern Bilgrau <abilgrau (at) math.aau.dk
}
\references{
See \url{http://brainarray.mbni.med.umich.edu}.
}
\seealso{
\code{\link{tempdir}}, \code{\link{install.packages}}
}

