% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cleanMetadata}
\alias{cleanMetadata}
\title{Generic metadata cleaning function}
\usage{
cleanMetadata(meta_data, ...)
}
\arguments{
\item{geo_nbr}{A \code{character} giving the GEO number.}
}
\value{
Returns a function that will clean the given GEO dataset.
}
\description{
S3 method dispatch giving the custom functions for formatting datasets.
}
\keyword{internal}

