% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanMetadata.GSE10846.R, R/cleanMetadata.GSE11318.R, R/cleanMetadata.GSE31312.R, R/cleanMetadata.GSE34171.R, R/cleanMetadata.R
\name{cleanMetadata.GSE10846}
\alias{cleanMetadata}
\alias{cleanMetadata.GSE10846}
\alias{cleanMetadata.GSE11318}
\alias{cleanMetadata.GSE31312}
\alias{cleanMetadata.GSE34171}
\alias{cleanMetadata.data.frame}
\title{Generic metadata cleaning function}
\usage{
\method{cleanMetadata}{GSE10846}(meta_data)

\method{cleanMetadata}{GSE11318}(meta_data)

\method{cleanMetadata}{GSE31312}(meta_data)

\method{cleanMetadata}{GSE34171}(meta_data)

cleanMetadata(meta_data)

\method{cleanMetadata}{data.frame}(meta_data)
}
\arguments{
\item{meta_data}{Output from \code{downloadAndPrepareMetadata}.}
}
\value{
Returns a specfically cleaned \code{data.frame} for the GEO dataset.
}
\description{
S3 method dispatch giving the custom functions for formatting datasets.
The function \code{downloadAndPrepareMetadata} adds a class \code{"GSEXXXX"}
of the GEO number in question.
The generic looks for functions of the name \code{cleanMetadata.GSEXXXX} and
specific for that dataset and cleans metadata using that function.
}
\details{
GSE10846:\cr
   The cleanup of GSE10846 (LLMPP) adds two batches corresponding to each
   the CHOP and the R-CHOP cohort.

GSE11318:\cr
   The cleanup of GSE11318 (NCI) adds the batch corresponding to
   platform GPL570 (HG-U133 plus 2) and excludes the GPL6400 samples.

GSE31312:\cr
   This function makes the samples GSM776068, GSM776149, and GSM776462 to be
   left out of the downsteam preprocessing due to bad array quality.

GSE34171:\cr
   The cleanup of GSE34171 (MDFCI) adds three batches corresponding to each
   platform (HG-U133 plus 2, HG-U133A, HG-U133B).
}
\author{
Anders Ellern Bilgrau, Steffen Falgreen Larsen
}

