% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DLBCLdata-package.R
\docType{data}
\name{DLBCL_overview}
\alias{DLBCL_overview}
\title{Overview of available DLBCL data}
\format{A \code{data.frame} giving information about the DLBCL studies with the
following columns:\cr
\describe{
  \item{GSE}{The accession number}
  \item{Study}{A name or abbreviation of the study.}
  \item{FullName}{The full name of the study.}
  \item{Author}{The first author of the study.}
  \item{ArrayTypes}{The array type(s) used for the gene expression profiles.}
  \item{Citation}{Citation information. Please double check the information
    if the data is used.}
}}
\description{
A \code{data.frame} of the manually checked and curated DLBCL datasets in
the \pkg{DLBCLdata}-package.
The \code{data.frame} gives information on the GEO number, a study
acronym, the expanded acronym, the principal author, the arraytypes
present in the study, and a full citation.
}
\examples{
data(DLBCL_overview)
print(DLBCL_overview[,-6])
#View(DLBCL_overview)
}
\keyword{data}
\keyword{datasets,}

