% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{DLBCLdata-package}
\alias{DLBCL}
\alias{DLBCLdata}
\alias{DLBCLdata-package}
\alias{dlbcl}
\alias{dlbcldata}
\title{DLBCLdata: Diffuse Large B-cell Lymphoma data}
\description{
An R-package for reproducible and easily available datasets in DLBCL.
}
\examples{
2+2 # FILL
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk> \cr
  Steffen Falgreen Larsen <sfl (at) rn.dk>
}

