% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DLBCLdata-package.R
\docType{package}
\name{DLBCLdata-package}
\alias{DLBCL}
\alias{DLBCLdata}
\alias{DLBCLdata-package}
\alias{dlbcl}
\alias{dlbcldata}
\title{\pkg{DLBCLdata}: Automated and reproducible download and preprocessing of
  DLBCL data}
\description{
An R-package for reproducible and easily available gene expression profile
(GEP) datasets in  Diffuse Large B-cell Lymphoma (DLBCL).
This package automatically downloads and preproceses the microarray data
using the manufacturer's or custom Brainarray chip definition files (CDF).

The most important user function is \code{\link{downloadAndProcessGEO}} and
the functions documented from there. For more information, read the README
file \href{https://github.com/AEBilgrau/DLBCLdata}{here at GitHub.}
}
\examples{
# Overview of the curated available data:
data(DLBCL_overview)
print(DLBCL_overview[,-6])

# Overview of all functions in DLBCLdata
ls("package:DLBCLdata")
}
\author{
Anders Ellern Bilgrau\cr
  Steffen Falgreen Larsen
}

