% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadAndPrepareGEOData}
\alias{downloadAndPrepareGEOData}
\title{Download GEO dataset}
\usage{
downloadAndPrepareGEOData(geo_nbr, destdir = getwd(), verbose = TRUE,
  clean = TRUE)
}
\arguments{
\item{geo_nbr}{The GEO ascession number.}

\item{destdir}{The destination dir of the downloaded files.}

\item{verbose}{Signal the process.}

\item{clean}{Should the strictly unnessesary files be deleted?}
}
\description{
Function for downloading a GEO dataset. Simply a wrapper for
and \code{getGEOSuppFiles} which downloads and unzips the
\code{.CEL} files in the GEO series.
}
\note{
The function will overwrite existing files in the \code{destdir}.
}
\examples{
\dontrun{
downloadAndPrepareGEOData(geo_nbr = "GSE18376",
                          destdir = tempdir())
}
}
\author{
Anders Ellern Bilgrau,
  Steffen Falgreen Larsen
}

