% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/downloadAndPrepareMetadata.R
\name{downloadAndPrepareMetadata}
\alias{downloadAndPrepareMetadata}
\title{Download GEO metadata}
\usage{
downloadAndPrepareMetadata(geo_nbr, destdir = getwd(), clean = FALSE,
  verbose = TRUE)
}
\arguments{
\item{geo_nbr}{The GEO ascession number.}

\item{destdir}{The destination dir of the downloaded files.}

\item{clean}{Should the strictly unnessesary files be deleted?}

\item{verbose}{Signal the process.}
}
\value{
A \code{list} of \code{data.frame}s giving the clinical and metadata
  information for the GEO dataset.
}
\description{
Function for downloading the metadata of a GEO dataset and prepare it for
analysis.
The function uses \code{getGEOSuppFiles} which downloads the data. The
function then unzips the data files in the GEO series.
}
\note{
The function will overwrite existing files in the \code{destdir}.
}
\examples{
\dontrun{
data(DLBCL_overview)
geo_nbr <- DLBCL_overview[6,1]
downloadAndPrepareMetadata(geo_nbr = geo_nbr, destdir = getwd())
}
}
\author{
Anders Ellern Bilgrau,
  Steffen Falgreen Larsen
}
\keyword{internal}

