% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/downloadAndProcessDLBCL.R
\name{downloadAndProcessDLBCL}
\alias{downloadAndProcessDLBCL}
\title{Download and process all curated DLBCL datasets}
\usage{
downloadAndProcessDLBCL(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link{downloadAndProcessGEO}}.}
}
\value{
Saves a \code{list} of the processed datasets in the working
  directory named "dlbcl_data.Rds".
  Invisibly returns a \code{list} of the output of
  \code{\link{downloadAndProcessGEO}}.
}
\description{
Automatically download and process all support curated DLBCL datasets.
Simply loops over all datasets in \code{data(dlbcl_overview)}.
}
\examples{
listTargets("brainarray")
\dontrun{
# Warning, very long processing times if data is not downloaded.

# Preprocess with brainarray "Entrez" gene ids
res <- downloadAndProcessDLBCL(cdf = "brainarray", target = "ENTREZG",
                               clean = FALSE)

# Preprocess with brainarray "Ensembl" gene ids
res2 <- downloadAndProcessDLBCL(cdf = "brainarray", target = "ENSG",
                                clean = TRUE)

# Preprocess with affy's probesets
res3 <- downloadAndProcessDLBCL()  #= downloadAndProcessDLBCL(cdf = "affy")
}
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk> \cr
  Steffen Falgreen Larsen <sfl (at) rn.dk>
}

