% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadAndProcessDLBCL}
\alias{downloadAndProcessDLBCL}
\title{Download all curated DLBCL datasets}
\usage{
downloadAndProcessDLBCL(...)
}
\description{
Automatically download all curated DLBCL datasets.
}
\examples{
\dontrun{
res <- downloadAndProcessDLBCL()
}
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
  Steffen Falgreen Larsen <sfl (at) rn.dk>
}

