% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadAndProcessGEO}
\alias{downloadAndProcessGEO}
\title{Download GEO dataset and preprocess it}
\usage{
downloadAndProcessGEO(geo_nbr, destdir = getwd(), ..., clean = FALSE,
  verbose = TRUE)
}
\arguments{
\item{geo_nbr}{The GEO ascession number.}

\item{destdir}{The destination dir of the downloaded files.}

\item{...}{Arguments passed to \code{\link{preprocessCELFiles}}}

\item{clean}{Should the strictly unnessesary files be deleted?}

\item{verbose}{Signal the process.}
}
\value{
Creates a folder with the downloaded and processed files.
  The processed files are saved as an \code{.Rds} object named
  after the used parameters.
  Invisibly returns the saved object.
}
\description{
Functionality for automatically downloading, RMA preprocessing the array
data and formatting the phenotype data, and saving the results in a
\code{.Rds} file.
}
\note{
The function will overwrite existing files in the \code{destdir}.
}
\examples{
\dontrun{
data(DLBCL_overview)
geo_nbr <-  DLBCL_overview[6,1]
res <- downloadAndProcessGEO(geo_nbr = geo_nbr, cdf = "brainarray",
                             target = "ENSG", clean = FALSE)
}
}
\author{
Anders Ellern Bilgrau,
  Steffen Falgreen Larsen
}

