% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadAndProcessGEO}
\alias{downloadAndProcessGEO}
\title{Download GEO dataset and preprocess it}
\usage{
downloadAndProcessGEO(geo_nbr, destdir = getwd(), ..., clean = FALSE,
  verbose = TRUE)
}
\arguments{
\item{geo_nbr}{The GEO ascession number.}

\item{destdir}{The destination dir of the downloaded files.}

\item{...}{Arguments passed to \code{\link{preprocessCELFiles}}}

\item{clean}{Should the strictly unnessesary files be deleted?}

\item{verbose}{Signal the process.}
}
\description{
Functionality for downloading a GEO dataset and preprocess it.
}
\note{
The function will overwrite existing files in the \code{destdir}.
}
\examples{
\dontrun{
downloadAndProcessGEO(geo_nbr = "GSE18376")
}
}
\author{
Anders Ellern Bilgrau,
  Steffen Falgreen Larsen
}

