% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLatestVersion.R
\name{getLatestVersion}
\alias{getLatestVersion}
\title{Get latest brainarray version number}
\usage{
getLatestVersion()
}
\value{
Returns a string on the form X.Y.Z giving the latest version.
}
\description{
Currently a very hacky implementation. To be updated later. The function
currently reads the tables available at
\url{http://brainarray.mbni.med.umich.edu/Brainarray/Database/CustomCDF/CDF_download.asp}
to determine the latest version
}
\examples{
getLatestVersion()
}
\keyword{internal}

