% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocessCELFiles}
\alias{preprocessCELFiles}
\title{RMA preprocess CEL files}
\usage{
preprocessCELFiles(cel_files = list.files(path, pattern = "\\\\CEL$"),
  cdf = "affy", target = c("core", "full", "feature"),
  version = getLatestVersion(), background = TRUE, normalize = TRUE,
  path = getwd())
}
\arguments{
\item{cel_files}{}

\item{cdf}{}

\item{target}{}

\item{version}{}

\item{normalize}{}

\item{backgroud}{}
}
\value{
An expression set object.
}
\description{
General function for RMA processing microarray data. Automatically downloads
custom Brainarray chip definition files (cdf) if wanted.
}

