% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocessCELFiles.R
\name{preprocessCELFiles}
\alias{preprocessCELFiles}
\title{RMA preprocess CEL files}
\usage{
preprocessCELFiles(cel_files, cdf = c("affy", "brainarray"), target,
  version = getLatestVersion(), background = TRUE, normalize = TRUE,
  path = getwd())
}
\arguments{
\item{cel_files}{A character vector of .CEL files.}

\item{cdf}{A character specifying the CDF type to use. Should be either
\code{"brainarray"} or \code{"affy"}.}

\item{target}{Specifies the "target" CDF to use depending on \code{cdf}.
If \code{cdf} is \code{"affy"} then \code{target} is e.g. \code{"core"}.
If \code{cdf} is \code{"brainarray"} then \code{target} is
e.g. \code{"ensg"}, \code{"entrez"}.
See \code{listTargets()} to list the available targest.}

\item{version}{A character giving the brainarray version to be used.
Only used if \code{cdf} is \code{"brainarray"}. See \code{listVersions()}.}

\item{background}{boolean. Should RMA background correction be performed?}

\item{normalize}{boolean. Should RMA normalization be performed?}
}
\value{
An expression set object.
}
\description{
General function for RMA processing microarray data. Automatically downloads
custom Brainarray chip definition files (cdf) if wanted.
}

