% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readDACData.R
\name{readDACData}
\alias{readDACData}
\title{Read DLBCL class files}
\usage{
readDACData(geo_nbr)
}
\arguments{
\item{geo_nbr}{A \code{character} string giving the GSE number.}
}
\value{
A \code{data.frame} of the predicted classes and the class
  probabilities.
}
\description{
Functions to read the internal data files of classified CEL files using
the DLBCL Automatic Classifier (DAC) [1]. These were manually supplied to the
DAC Windows interface.
}
\examples{
head(readDACData("GSE19246"))
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk> \cr
  Steffen Falgreen
}
\references{
[1] Care, M. A., Barrans, S., Worrillow, L., Jack, A., Westhead, D. R., &
      Tooze, R. M. (2013). A microarray platform-independent classification
      tool for cell of origin class allows comparative analysis of gene
      expression in diffuse large B-cell lymphoma.
      PloS One, 8(2), e55895. doi:10.1371/journal.pone.0055895
}

