% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readDACData.R
\name{readDACData}
\alias{readDACData}
\title{Read DLBCL class files}
\usage{
readDACData(geo_nbr)
}
\arguments{
\item{geo_nbr}{A \code{character} string giving the GSE number.}
}
\value{
A \code{data.frame} of the predicted classes and the class
  probabilities.
}
\description{
Functions to read the internal data files of classified CEL files using
the DLBCL Automatic Classifier (DAC). These were manually supplied to the
DAC Windows interface.
}
\examples{
head(readDACData("GSE19246"))
}
\keyword{internal}

